requirements.txt
streamlit
pydantic
sqlalchemy

langchain
langgraph
langchain-openai

openai

pyswisseph
reportlab

passlib[bcrypt]
python-dotenv
requests
tiktoken
typing-extensions